pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "cloudrunfastapi"
authors = [{name = "Anthony Corletti", email = "anthcor@gmail.com"}]
readme = "README.md"
classifiers = ["License :: OSI Approved :: MIT License"]
dynamic = ["version", "description"]
dependencies = [
    "fastapi>=0.70.1",
    "gunicorn>=20.1.0",
    "uvicorn>=0.16.0",
    "SQLAlchemy>=1.4.22",
    "SQLAlchemy-Utils>=0.37.8",
    "psycopg2-binary>=2.9.1",
    "alembic>=1.6.5",
    "cryptography>=3.4.7",
    "authlib>=0.15.4",
    "PyJWT>=2.1.0",
    "passlib>=1.7.4",
    "email-validator>=1.1.3",
    "python-multipart>=0.0.5",
    "bcrypt>=3.2.0"
]

[project.optional-dependencies]
test = [
    "pytest>=6.2.5",
    "mypy>=0.910",
    "coverage>=6.1.1",
    "pytest-cov>=3.0.0",
]
dev = [
    "flake8>=3.9.2",
    "black>=21.10b0",
    "isort>=5.9.3",
    "autoflake>=1.4",
    "flake8-docstrings>=1.6.0",
]

[project.urls]
Home = "https://github.com/anthonycorletti/cloudrun-fastapi"

[tool.isort]
profile = "black"
